+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   - Fixed a bug with mocking dynamic parameters on latest Windows 10 / PSv5 builds.  [GH-419]
   - Fix for NUnit XML export on .NET core.  [GH-420]
   - Mocking improvements for calling original commands with begin/process/end blocks. [GH-422]
+  - Case insensitive replacement of Test in help [GH-428]
+  - Improve stack trace and exception console output [GH-426]
 
 ## 3.3.11 (September 8, 2015)
   - Fixed a bug where mocking New-Object would cause a stack overflow.  [GH-405]